mod-comments: move _ur_cli into _ur directory as cli-scripts
